requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.12.0
google-generativeai==0.8.3
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
googlemaps==4.10.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
tenacity==9.0.0
reportlab==4.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
